ESP32: return size_t from HardwareCapabilities heap getters
heap_caps_get_*_size() returns size_t, and existing log call sites format the
internal free heap with %u. Returning uint32_t (long unsigned on xtensa) broke
-Werror=format in comms_ip.cpp safe_send() on the CI toolchain.

+++ firmware/adsbee_1090/esp/main/hardware_capabilities.cpp
@@ -9,7 +9,7 @@
 #endif
 
 bool HardwareCapabilities::has_psram_ = false;
-uint32_t HardwareCapabilities::psram_total_bytes_ = 0;
+size_t HardwareCapabilities::psram_total_bytes_ = 0;
 
 void HardwareCapabilities::Detect() {
 #ifdef CONFIG_SPIRAM
